server/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.3
flask-cors==4.0.0 